requirements.txt
Flask
opencv-python
face_recognition
numpy
pandas
openpyxl
Pillow
